Rainbow complete
I got the rainbow effect done for the mask. I still have to define how many leds are in the string, but other then that I should be on my way to completing most of the mask. I still need to code the main glitch effect that will make this all come together. I will be updating that in two parts do to the fact that I currently do not have the full string put together. -GJ